refactor(portfolio): extract factories into PortfolioConstructorMixin
Move from_risk_position/from_position and the _evaluate_position_expr
helper into _portfolio_constructors.py. Both factories now funnel through
cls.from_cash_position (which stays on the dataclass), so no raw cls(...)
construction happens in a mixin and mypy --strict stays happy. Behaviour
is identical; from_cash_position with a DataFrame position is the same
path the old code reached.

portfolio.py: 793 → 451 lines.
